feat: implementing JWT

+++ src/main/java/com/game/pokedex/controllers/UserController.java
@@ -5,9 +5,14 @@ import com.game.pokedex.dtos.UserRequest;
 import com.game.pokedex.dtos.UserUpdateRequest;
 import com.game.pokedex.service.UserService;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+import org.springframework.security.core.Authentication;
 
 import java.util.List;
+
+import static com.game.pokedex.config.PermissionValitation.validatePermission;
 
 @RestController
 @RequestMapping("/user")
@@ -19,12 +24,16 @@ public class UserController {
     }
 
     @GetMapping
+    @PreAuthorize("hasRole(T(com.game.pokedex.entities.User.Role).ADMIN.name())")
     public List<UserDto> getUsers(){
         return this.userService.getAll();
     }
 
     @DeleteMapping("/{username}")
-    public void deleteUserByUsername(@PathVariable("username") String username) throws Exception {
+    public void deleteUserByUsername(@PathVariable("username") String username, Authentication authentication) throws Exception {
+        if (validatePermission.apply(authentication, username)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Acesso negado");
+        }
         this.userService.deleteUserByUsername(username);
     }
